Pass a full URI to the HTTP request factory instead of having to figure it out from the API base URL, version path, and requested path
The goal is to bring the API base URL closer to the client, so that we can inject it instead of relying on the DNSimple global state
