+++ src/main/java/com/dnsimple/endpoints/http/HttpRequestFactory.java
@@ -2,6 +2,7 @@ package com.dnsimple.endpoints.http;
 
 import com.dnsimple.Dnsimple;
 
+import java.net.URI;
 import java.util.Map;
 
 /**
@@ -24,5 +25,5 @@ public interface HttpRequestFactory {
      * @param body              the HTTP request's body payload
      * @return an RawResponse object with the status code, headers and body of the HTTP response
      */
-    RawResponse execute(String userAgent, String accessToken, HttpMethod method, String path, Map<String, Object> queryStringParams, Object body);
+    RawResponse execute(String userAgent, String accessToken, HttpMethod method, URI uri, Map<String, Object> queryStringParams, Object body);
 }
